Backported changes from the development cvs version. Those are: - Added gnutls_certificate_set_params_function() and   gnutls_anon_set_params_function() that set the RSA or DH   parameters using a callback. - Added functions gnutls_rsa_params_cpy(), gnutls_dh_params_cpy()   and gnutls_x509_privkey_cpy(). - Corrected a compilation issue when opencdk was installed in a   non standard directory.

+++ libextra/openpgp/xml.c
@@ -140,7 +140,7 @@ xml_add_key( gnutls_string *xmlkey, int ext, cdk_pkt_pubkey_t pk, int sub )
     const char *algo, *s;
     char keyid[16], fpr[41], tmp[32];
     uint8 fingerpr[20];
-    unsigned long kid[2];
+    unsigned int kid[2];
     int i = 0, rc = 0;
 
     if( !xmlkey || !pk ) {
@@ -254,7 +254,7 @@ xml_add_sig( gnutls_string *xmlkey, int ext, cdk_pkt_signature_t sig )
 {
     const char *algo, *s;
     char tmp[32], keyid[16];
-    unsigned long kid[2];
+    unsigned int kid[2];
     int rc = 0;
 
     if( !xmlkey || !sig ) {
